+++ src/msdb5/card/card_number_test.go
@@ -18,7 +18,7 @@ func TestId40IsKing(t *testing.T) {
 	verifyCorrectNumber(t, 40, 10)
 }
 
-func verifyCorrectNumber(t *testing.T, id, number int) {
+func verifyCorrectNumber(t *testing.T, id, number uint8) {
 	card, _ := ByID(id)
 	if card.number != uint8(number) {
 		t.Fatalf("Card %v's number is not %d", card, number)
